storeregi update01

+++ src/com/yameokja/mc/StoreregiDTO.java
@@ -3,7 +3,7 @@ package com.yameokja.mc;
 public class StoreregiDTO
 {
 	private String user_num, st_in_file, st_place_file, st_name, st_location, st_location_dt, st_tel, admin_read_date, reg_date;
-	private int st_in_num, st_place_num;
+	private long st_in_num, st_place_num;
 	
 	// getter / setter
 	public String getUser_num()
@@ -78,22 +78,23 @@ public class StoreregiDTO
 	{
 		this.reg_date = reg_date;
 	}
-	public int getSt_in_num()
+	public long getSt_in_num()
 	{
 		return st_in_num;
 	}
-	public void setSt_in_num(int st_in_num)
+	public void setSt_in_num(long st_in_num)
 	{
 		this.st_in_num = st_in_num;
 	}
-	public int getSt_place_num()
+	public long getSt_place_num()
 	{
 		return st_place_num;
 	}
-	public void setSt_place_num(int st_place_num)
+	public void setSt_place_num(long st_place_num)
 	{
 		this.st_place_num = st_place_num;
 	}
+
 	
 	
 }
